refactor: revert test changes
