+++ packages/preset-mini/src/index.ts
@@ -1,4 +1,6 @@
 import type { Postprocessor, Preflight, PreflightContext, Preset, PresetOptions } from '@unocss/core'
+import { extractorSplit } from '@unocss/core'
+import { extractorArbitraryVariants } from '@unocss/extractor-arbitrary-variants'
 import { preflights } from './preflights'
 import { rules } from './rules'
 import type { Theme, ThemeAnimation } from './theme'
@@ -58,6 +60,15 @@ export interface PresetMiniOptions extends PresetOptions {
    * @default true
    */
   preflight?: boolean
+
+  /**
+   * Enable arbitrary variants, for example `<div class="[&>*]:m-1 [&[open]]:p-2"></div>`.
+   *
+   * Disable this might slightly improve the performance.
+   *
+   * @default true
+   */
+  arbitraryVariants?: boolean
 }
 
 export function presetMini(options: PresetMiniOptions = {}): Preset<Theme> {
@@ -72,9 +83,19 @@ export function presetMini(options: PresetMiniOptions = {}): Preset<Theme> {
     rules,
     variants: variants(options),
     options,
-    postprocess: VarPrefixPostprocessor(options.variablePrefix),
-    preflights: options.preflight ? normalizePreflights(preflights, options.variablePrefix) : [],
     prefix: options.prefix,
+    postprocess: VarPrefixPostprocessor(options.variablePrefix),
+    preflights: options.preflight
+      ? normalizePreflights(preflights, options.variablePrefix)
+      : [],
+    configResolved(config) {
+      if (options.arbitraryVariants === false)
+        return
+      if (config.extractors.includes(extractorSplit))
+        config.extractors.splice(config.extractors.indexOf(extractorSplit), 1, extractorArbitraryVariants)
+      else
+        config.extractors.unshift(extractorArbitraryVariants)
+    },
   }
 }
