+++ packages/preset-wind/package.json
@@ -40,7 +40,6 @@
   },
   "dependencies": {
     "@unocss/core": "workspace:*",
-    "@unocss/extractor-arbitrary-variants": "workspace:*",
     "@unocss/preset-mini": "workspace:*"
   }
 }
