+++ test/pos.test.ts
@@ -6,7 +6,6 @@ import { createGenerator, extractorSplit } from '@unocss/core'
 import { getMatchedPositionsFromCode as match } from '@unocss/shared-common'
 import transformerVariantGroup from '@unocss/transformer-variant-group'
 import cssDirectives from '@unocss/transformer-directives'
-import presetArbitrary from '@unocss/extractor-arbitrary-variants'
 import extractorPug from '@unocss/extractor-pug'
 
 describe('matched-positions', async () => {
@@ -15,7 +14,6 @@ describe('matched-positions', async () => {
       presets: [
         presetAttributify({ strict: true }),
         presetUno({ attributifyPseudo: true }),
-        presetArbitrary(),
       ],
     })
 
@@ -41,21 +39,6 @@ describe('matched-positions', async () => {
             21,
             22,
             "[border=\\"2\\"]",
-          ],
-          [
-            23,
-            37,
-            "[border=\\"[&_span]:white\\"]",
-          ],
-          [
-            46,
-            65,
-            "[&>span]:text-white",
-          ],
-          [
-            46,
-            65,
-            "[hover=\\"[&>span]:text-white\\"]",
           ],
           [
             67,
@@ -86,6 +69,11 @@ describe('matched-positions', async () => {
             13,
             15,
             "[border=\\"bb\\"]",
+          ],
+          [
+            16,
+            17,
+            "b",
           ],
           [
             16,
@@ -130,7 +118,6 @@ describe('matched-positions', async () => {
     const uno = createGenerator({
       presets: [
         presetUno(),
-        presetArbitrary(),
       ],
       shortcuts: {
         '<custom-shortcut': 'text-lg',
@@ -156,11 +143,6 @@ describe('matched-positions', async () => {
             "hover:scale-100",
           ],
           [
-            44,
-            63,
-            "[&>span]:text-white",
-          ],
-          [
             64,
             80,
             "<custom-shortcut",
@@ -173,7 +155,6 @@ describe('matched-positions', async () => {
     const uno = createGenerator({
       presets: [
         presetUno(),
-        presetArbitrary(),
       ],
     })
